Added a fix for the issue with multiple listeners being created
Earlier there was an issue with the event listeners in the Cy API calls 
section of the cytoscape object, where the listeners were being created 
over and over again - fixed this by adding a 'initialised state' which 
updates at the end of that block, and then if it's true it wont 
re-create the listeners

Currently still having trouble with the selected nodes to make the egde. 
Seems the value in the cy prop always stays at '' and never updates.